Fixed Intake to reflect new Intake changes

+++ src/org/usfirst/frc/team1360/robot/subsystem/IntakeProvider.java
@@ -2,7 +2,7 @@ package org.usfirst.frc.team1360.robot.subsystem;
 
 public interface IntakeProvider{
 	
-	void setPosition(int position);
+	void setPosition(double position);
 	
 	int getPosition();
